arena.c: implement heap chunk defragmentation
If all memory is allocated and then freed, the user could request
a block of memory that is bigger than the largest chunk of memory.
the allocator does have the memory but do to heap fragmentation is
unable to find a chunk big enough to hold the users request.

This and the previous patch introduce the nessicary logic to handle
heap defragmentation.